WIP. Added data logger to primary project. Need to add enable/disable file logging.

+++ Force_Test_Stand_Update/HX711.cpp
@@ -234,7 +234,7 @@ long HX711::read_average(byte times) {
 	return sum / times;
 }
 
-double HX711::get_value_fast() {
+long HX711::get_value_fast() {
 	return read() - OFFSET;
 }
